testing_interface/requirements.txt
# Testing Interface Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
websockets==12.0

# Database
supabase==2.0.2
asyncpg==0.29.0

# AI
google-generativeai==0.3.2

# Data handling
pydantic==2.5.2
pandas==2.1.4

# Static files
aiofiles==23.2.1

# Date handling
python-dateutil==2.8.2